Matrix/requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.4
distro==1.8.0
Django==4.1.4
djangorestframework==3.14.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-spectacular==0.25.0
idna==3.4
inflection==0.5.1
jsonschema==3.2.0
numpy==1.23.5
packaging==22.0
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==5.4.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
texttable==1.6.7
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==0.59.0
